Fix bug with multiplication.
Using multiplication of a numpy data type with a state
did not result in __rmul__ being called on state.

The fix here is to not support  len on state (same as state.basis_size).
This leads to __mul__ called on the other data type failing.

Additionally add and extend docstrings.